Flow field addition
I made a flow field with the help of the lecture that showed flow fields. I then added a color palette, life to the lines, and made the lines more ribbon like. I'm not happy with the result just yet, I need to fix the ribbons so that they are nicer, and make the flow field more consistent so that all of the lines don't start at tha same point...